+++ base/pad_entry.c
@@ -18,6 +18,7 @@
  */
 
 #include "pad_entry.h"
+#include "util.h"
 
 pad_entry_node* init_pad_node = NULL;
 char *entry_id;
@@ -27,34 +28,7 @@ char auth_protocol[50];
 char auth_method[40];
 char ssecret[40];
 
-void remove_all_chars(char* str, char c) {
-    char *pr = str, *pw = str;
-    while (*pr) {
-        *pw = *pr++;
-        pw += (*pw != c);
-    }
-    *pw = '\0';
-}
 
-
-void strupp(char* beg)
-{
-    while (*beg = toupper(*beg)) beg++;
-}
-
-int hex_to_int(char c){
-        int first = c / 16 - 3;
-        int second = c % 16;
-        int result = first*10 + second;
-        if(result > 9) result--;
-        return result;
-}
- 
-int hex_to_ascii(char c, char d){
-        int high = hex_to_int(c) * 16;
-        int low = hex_to_int(d);
-        return high+low;
-}
 
 pad_entry_node* getPADEntry(char* remote_ts){
