+++ base/utils.c
@@ -497,6 +497,33 @@ sock_ntop(const struct sockaddr *sa, socklen_t salen) {
 }
 
 
+void remove_all_chars(char* str, char c) {
+    char *pr = str, *pw = str;
+    while (*pr) {
+        *pw = *pr++;
+        pw += (*pw != c);
+    }
+    *pw = '\0';
+}
 
+
+void strupp(char* beg)
+{
+    while (*beg = toupper(*beg)) beg++;
+}
+
+int hex_to_int(char c){
+        int first = c / 16 - 3;
+        int second = c % 16;
+        int result = first*10 + second;
+        if(result > 9) result--;
+        return result;
+}
+ 
+int hex_to_ascii(char c, char d){
+        int high = hex_to_int(c) * 16;
+        int low = hex_to_int(d);
+        return high+low;
+}
