+++ base/utils.h
@@ -124,7 +124,14 @@ void address_print(struct sadb_ext *ext);
 void key_print(struct sadb_ext *ext);
 
 
+#ifndef UTIL_H
+#define UTIL_H
+void remove_all_chars(char*, char);
+void strupp(char*);
+void hex_to_int(char);
+void hex_to_ascii(char, char);
 
+#endif
